Fish move away from player

+++ src/Movable.java
@@ -17,7 +17,7 @@ public class Movable {
         this.speed = speed;
     }
 
-    public double distanceTo(int mx, int my) {
+    public double distanceTo(double mx, double my) {
         return Math.sqrt((mx - x) * (mx - x) + (my - y) * ((my - y)));
     }
